% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{alluvial}
\alias{alluvial}
\title{Alluvial Aquifer}
\format{A tibble with 59 rows and 2 variables:
\describe{
  \item{\code{Median_grain_size}}{median grain size particle, in mm}
  \item{\code{Yield}}{groundwater yield, in \eqn{gal/day/ft^{2}}}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
alluvial
}
\description{
Grain size and permeability of alluvial aquifers
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
