% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{atrazine}
\alias{atrazine}
\title{Atrazine Concentration}
\format{A tibble with 48 rows and 2 variables:
\describe{
  \item{\code{Month}}{month of atrazine concentration measurement}
  \item{\code{Concentration}}{atrazine concentration, in ppm}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
atrazine
}
\description{
Atrazine concentration in shallow groundwater before (June) and after (September) the application season.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
