% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{conecuh_river}
\alias{conecuh_river}
\title{Conecuh River Streamflows}
\format{A tibble with 20 rows and 2 variables:
\describe{
  \item{\code{Year}}{year of streamflows recording}
  \item{\code{Flow}}{annual streamflows, in cfs}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
conecuh_river
}
\description{
Annual streamflows for the Conecuh River, AL data.
}
\note{
The data come from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
