% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{contamination}
\alias{contamination}
\title{Contamination Rating}
\format{A tibble with 124 rows and 5 variables:
\describe{
  \item{\code{UZT}}{unsaturated zone thickness, possible scores: 0 (favorable) to 9 (unfavorable)}
  \item{\code{AY}}{yields: aquifer properties, possible scores: 0 (poor) to 6 (good)}
  \item{\code{GWQ}}{groundwater quality, possible scores: 0 (poor) to 5 (excellent)}
  \item{\code{HWM}}{hazard rating for source, possible scores: 1 (low) to 9 (high)}
  \item{\code{CR}}{contamination rating, possible scores: 0 (contaminated) to 1 (not contaminated)}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
contamination
}
\description{
Factors affecting contamination from impoundments
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
