% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{corbicula}
\alias{corbicula}
\title{Corbicula}
\format{A tibble with 108 rows and 4 variables:
\describe{
  \item{\code{Year}}{year of measurement}
  \item{\code{Season}}{season type}
  \item{\code{Strata}}{strata in the Tennessee River}
  \item{\code{Corbicula}}{the number of Corbicula per square meter}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
corbicula
}
\description{
Corbicula on the Tennessee River
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
