% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{iron}
\alias{iron}
\title{Iron Concentration}
\format{A tibble with 78 rows and 3 variables:
\describe{
  \item{\code{Rock}}{rock type}
  \item{\code{Mine}}{mining history}
  \item{\code{Concentration}}{iron concentration, in ppm}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
iron
}
\description{
Subset of iron concentrations at low flow in numerous small stearms in the coal-producing areas of eastern Ohio.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
