% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{phosphorus}
\alias{phosphorus}
\title{Phosphorus Transport}
\format{A tibble with 96 rows and 4 variables:
\describe{
  \item{\code{Year}}{year of measurement}
  \item{\code{Conc}}{phosphorus concentration, in ppm}
  \item{\code{Load}}{Conc X Q, in ton/day}
  \item{\code{Q}}{stream discharge, in cfs}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
phosphorus
}
\description{
Phosphorus transport, Illinois River at Marseilles
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
