% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{potomac_river}
\alias{potomac_river}
\title{Potomac River Streamflows}
\format{A tibble with 365 rows and 2 variables:
\describe{
  \item{\code{Year}}{year of streamflows measurement}
  \item{\code{Flow}}{daily streamflows, in cfs}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
potomac_river
}
\description{
Ranked daily streamflow for the Potomac River, Wash. D.C data.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
