% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{roe}
\alias{roe}
\title{ROE and TDS Data}
\format{A tibble with 84 rows and 3 variables:
\describe{
  \item{\code{ROE}}{residue on evaporation}
  \item{\code{TDS}}{total dissolved solid, in ppm}
  \item{\code{Date}}{date of measurement}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
roe
}
\description{
ROE and TDS data, Rappahannock R. near Fredricksburg, VA.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
