% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{runbas}
\alias{runbas}
\title{Runoff and Basin Characteristic}
\format{A tibble with 13 rows and 10 variables:
\describe{
  \item{\code{ROFF}}{mean annual runoff}
  \item{\code{PRECIP}}{precipitation}
  \item{\code{AREA}}{drainage area}
  \item{\code{SLOPE}}{average slope}
  \item{\code{LENGTH}}{the length of the drainage basin}
  \item{\code{PERIM}}{the perimeter of the basin}
  \item{\code{DI}}{the diameter of the largest circle which could be inscribed within the drainage basin }
  \item{\code{Rs}}{shape factor}
  \item{\code{FREQ}}{the stream frequency -- the ratio of the number of streams in the basin to the basin area}
  \item{\code{Rr}}{ relief ratio}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
runbas
}
\description{
Mean annual runoff and basin characteristics
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
