% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{saddle_river}
\alias{saddle_river}
\title{Saddle River Discharge}
\format{A tibble with 65 rows and 2 variables:
\describe{
  \item{\code{Year}}{year of discharges measurement}
  \item{\code{Flow}}{annual peak discharges, in cfs}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
saddle_river
}
\description{
Annual peak discharges for the Saddle River, NJ data.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
