% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{scwell}
\alias{scwell}
\title{Specific Capacity of Wells}
\format{A tibble with 200 rows and 2 variables:
\describe{
  \item{\code{Rock}}{rock type}
  \item{\code{Specific_Capacity}}{specific capacity, in gal/min/ft}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
scwell
}
\description{
Specific capacities of wells in Pennsylvania.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
