% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sediment}
\alias{sediment}
\title{Sediment Transport}
\format{A tibble with 21 rows and 3 variables:
\describe{
  \item{\code{Year}}{year of measurement}
  \item{\code{Load}}{load of sediment}
  \item{\code{Q}}{stream discharge}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
sediment
}
\description{
Green River, Kentucky sediment transport data.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
