% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sf_record}
\alias{sf_record}
\title{Streamflow Record}
\format{A tibble with 50 rows and 4 variables:
\describe{
  \item{\code{Year}}{year of measurement}
  \item{\code{Base_sf}}{base streamflow, in million cubic meters per month}
  \item{\code{Short_sf}}{short streamflow, in million cubic meters per month}
  \item{\code{Actual_early_sf}}{actual early streamflow at short, in million cubic meters per month}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
sf_record
}
\description{
Streamflow data used for record extension
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
