% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{tds}
\alias{tds}
\title{Total Disolved Solid}
\format{A tibble with 80 rows and 4 variables:
\describe{
  \item{\code{Month}}{month of measurement}
  \item{\code{Year}}{year of measurement}
  \item{\code{TDS}}{tds concentration, in ppm}
  \item{\code{Q}}{river discharge, in cfs}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
tds
}
\description{
TDS concentrations for the Cuyahoga River, Ohio
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
