% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{totalN}
\alias{totalN}
\title{Urban Total Nitrogen Loads}
\format{A tibble with 42 rows and 9 variables:
\describe{
  \item{\code{LOGTN}}{log total nitrogen load}
  \item{\code{LOGDA}}{log contributing area}
  \item{\code{LOGIMP}}{log impervious area}
  \item{\code{MJTEMP}}{mean minimum January temperature}
  \item{\code{MSRAIN}}{mean seasonal rainfall}
  \item{\code{PRES}}{percentage of area residential}
  \item{\code{PNON}}{percentage of area non-urban}
  \item{\code{PCOMM}}{percentage of area commercial}
  \item{\code{PIND}}{percentage of area industrial}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
totalN
}
\description{
Data from 42 small urban drainage basins located in several cities around United States.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
