% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{totalP}
\alias{totalP}
\title{Total P Trends Data}
\format{A tibble with 133 rows and 3 variables:
\describe{
  \item{\code{Date}}{date of measurement}
  \item{\code{Discharge}}{stream discharge, in 1000 cfs}
  \item{\code{Load}}{total P load, in tons/day}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
totalP
}
\description{
Maumee River, Ohio total P trends data
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
