% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{utds}
\alias{utds}
\title{Uranium and TDS in Groundwaters}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 3 columns.}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
utds
}
\description{
Uranium and TDS concentration in Ogallala aquifer.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
