% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{water_level}
\alias{water_level}
\title{Water Levels Data}
\format{A tibble with 90 rows and 2 variables:
\describe{
  \item{\code{Date}}{date of measurement}
  \item{\code{Elevation}}{elevation, in ft}
}}
\source{
\url{http://water.usgs.gov/pubs/twri/twri4a3/}
}
\usage{
water_level
}
\description{
Water levels, P-R-M system middle aquifer, NJ.
}
\note{
The data comes from Statistical Methods for Water Resources Book written by:
DR.Helsel and R.M. Hirsch (2002).
}
\keyword{datasets}
